src/requirements.txt
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
numpy
glfw
PyQt5